refactor(frontend): move toast helper out of hooks/
hooks/use-toast.ts was not a React hook. It was a plain function that
wraps sonner with our variant enum. Putting it under hooks/ violates
the only actual rule of that folder (must use the use* React-hook
convention) and made every import look like a hook call, which misled
readers.

Move it to lib/toast.ts where plain helper utilities live (sanitize,
cache, supabase, youtubeUrl, etc.) and update every caller. Behaviour
is unchanged — same exported function, same signature.

Made-with: Cursor
