+++ frontend/src/components/assignment/AssignmentEditor.tsx
@@ -4,7 +4,7 @@ import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { coursesService } from "@/services/courses"
-import { toast } from "@/hooks/use-toast"
+import { toast } from "@/lib/toast"
 import type { Assignment, AssignmentSubmission } from "@/types"
 import {
   FileText,
