+++ frontend/src/components/course/CourseReviews.tsx
@@ -3,7 +3,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Button } from "@/components/ui/button"
 import { coursesService } from "@/services/courses"
 import { useAuth } from "@/context/useAuth"
-import { toast } from "@/hooks/use-toast"
+import { toast } from "@/lib/toast"
 import type { CourseReview } from "@/types"
 import { Star, Trash2, MessageSquare } from "lucide-react"
 import PageSpinner from "@/components/ui/PageSpinner"
