+++ frontend/src/components/quiz/QuizEditor.tsx
@@ -4,7 +4,7 @@ import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { coursesService } from "@/services/courses"
-import { toast } from "@/hooks/use-toast"
+import { toast } from "@/lib/toast"
 import type { Quiz, QuizQuestionType } from "@/types"
 import { Plus, Trash2, Save, ClipboardList, Loader2, ArrowUp, ArrowDown, GraduationCap } from "lucide-react"
 import { useConfirm } from "@/components/ui/alert-dialog"
