+++ frontend/src/pages/Teacher/ModuleEditor.tsx
@@ -8,7 +8,7 @@ import { Badge } from "@/components/ui/badge"
 import { coursesService } from "@/services/courses"
 import { getErrorDetail } from "@/lib/errorDetail"
 import type { Module, Chapter } from "@/types"
-import { toast } from "@/hooks/use-toast"
+import { toast } from "@/lib/toast"
 import { moduleSchema, chapterSchema } from "@/lib/validations/course"
 import { useConfirm } from "@/components/ui/alert-dialog"
 import { Label } from "@/components/ui/label"
