+++ frontend/src/pages/Teacher/TeacherDashboard.tsx
@@ -10,7 +10,7 @@ import { Badge } from "@/components/ui/badge"
 import { coursesService } from "@/services/courses"
 import { courseSchema, type CourseFormData } from "@/lib/validations/course"
 import type { Course, Certificate } from "@/types"
-import { toast } from "@/hooks/use-toast"
+import { toast } from "@/lib/toast"
 import { Plus, Pencil, Trash2, BookOpen, Layers, BarChart3, Eye, EyeOff, ClipboardList, Users, Clock, CheckCircle, XCircle, Award, Search, Copy, RotateCcw, Archive } from "lucide-react"
 import { getErrorDetail } from "@/lib/errorDetail"
 import { toProxyImage } from "@/lib/images"
